# backend/docs/openapi.apifox.cn.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Light OSS API",
    "version": "1.0.0",
    "description": "供 APIFOX 导入的 OpenAPI 3.0 JSON。文档中的路径与 HTTP 方法集合与后端实际注册的 Gin 路由保持一致。"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地后端默认地址"
    }
  ],
...
  "paths": {
...
    "/api/v1/buckets/{bucket}/folders": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": ["文件夹"],
        "operationId": "listFolders",
        "summary": "列出完整的文件夹树",
        "description": "返回存储桶的所有已知文件夹节点，包括通过文件夹标记对象创建的空文件夹。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "文件夹树",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderNodeListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称无效。错误代码： invalid_bucket_name。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出文件夹。错误代码： folder_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["文件夹"],
        "operationId": "createFolder",
        "summary": "创建空文件夹",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "文件夹已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderNodeEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求体或文件夹输入无效。错误代码：invalid_request、invalid_bucket_name、invalid_prefix、invalid_folder_name。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到存储桶或父文件夹。错误代码： bucket_not_found, folder_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "文件夹已存在。错误代码： folder_exists。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "文件夹创建失败。错误代码： bucket_lookup_failed、folder_lookup_failed、object_store_failed、object_metadata_failed]。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["文件夹"],
        "operationId": "deleteFolder",
        "summary": "删除文件夹。",
        "description": "当 `recursive` 省略或为 `false` 时，只能删除空文件夹。当 `recursive` 为 `true` 时，文件夹以及其所有嵌套文件和子文件夹都会被软删除。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/FolderPathQuery"
          },
          {
            "$ref": "#/components/parameters/RecursiveQuery"
          }
        ],
        "responses": {
          "204": {
            "description": "文件夹已删除。无响应正文。"
          },
          "400": {
            "description": "存储桶名称、文件夹路径或递归查询无效。错误代码： invalid_bucket_name、invalid_folder_path、invalid_request、invalid_prefix。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "找不到文件夹。错误代码： folder_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "文件夹不为空，且未启用递归删除。错误代码：folder_not_empty。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "文件夹删除失败。错误代码： folder_lookup_failed ， folder_delete_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/folders/archive": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": ["文件夹"],
        "operationId": "downloadFolderArchive",
        "summary": "将文件夹下载为 ZIP 压缩包",
        "description": "为指定的非根目录文件夹路径流式返回 ZIP 压缩包。压缩包会保留当前文件夹名称作为顶层目录，包含所有嵌套文件，保留通过文件夹标记对象创建的空子文件夹，并排除内部 .light-oss-folder 标记文件。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/FolderPathQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP 压缩包流。",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称或文件夹路径无效。错误代码： invalid_bucket_name、invalid_folder_path、invalid_prefix。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "找不到文件夹。错误代码： folder_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "文件夹归档准备失败。错误代码： folder_lookup_failed、folder_archive_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/entries": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": ["文件夹"],
        "operationId": "listExplorerEntries",
        "summary": "在文件夹前缀下列出直接条目",
        "description": "条目先按目录排序，然后按文件名排序。next_cursor是相同前缀/search组合的不透明光标。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PrefixQuery"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/ExplorerLimitQuery"
          },
          {
            "$ref": "#/components/parameters/CursorQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "资源管理器条目页面。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplorerEntryListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称、前缀或光标无效。错误代码： invalid_bucket_name、invalid_prefix、invalid_cursor。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出资源管理器条目。错误代码： explorer_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/entries/batch-delete": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "post": {
        "tags": ["对象"],
        "operationId": "deleteExplorerEntriesBatch",
        "summary": "批量删除资源管理器条目",
        "description": "一次最多删除 200 个文件或目录。目录会递归删除，单项失败会返回在成功响应结果中。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items"],
                "properties": {
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 200,
                    "items": {
                      "type": "object",
                      "required": ["type", "path"],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["file", "directory"]
                        },
                        "path": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 512,
                          "description": "文件使用对象键；目录使用以 '/' 结尾的规范化文件夹路径。"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "包含 deleted_count、failed_count 和 failed_items 的批量处理结果。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称、请求体、条目数量、类型或路径无效。错误代码：invalid_request、invalid_bucket_name。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "存储桶不存在。错误代码：bucket_not_found。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buckets/{bucket}/objects": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BucketParam"
        }
      ],
      "get": {
        "tags": ["对象"],
        "operationId": "listObjects",
        "summary": "列表对象",
        "description": "对象首先返回最新。next_cursor是相同前缀过滤器的不透明光标。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PrefixQuery"
          },
          {
            "$ref": "#/components/parameters/ObjectListLimitQuery"
          },
          {
            "$ref": "#/components/parameters/CursorQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "对象页面。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶名称、前缀或光标无效。错误代码： invalid_bucket_name、invalid_prefix、invalid_cursor。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出对象。错误代码： object_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recycle-bin/objects": {
      "get": {
        "tags": ["回收站"],
        "operationId": "listRecycleBinObjects",
        "summary": "列出回收站条目",
        "description": "使用游标分页列出回收站中的逻辑文件和目录条目。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "description": "可选的存储桶名称筛选。",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/ObjectListLimitQuery"
          },
          {
            "$ref": "#/components/parameters/CursorQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "回收站条目和 next_cursor。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "存储桶筛选或游标无效。错误代码：invalid_bucket_name、invalid_cursor。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "无法列出回收站条目。错误代码：recycle_bin_list_failed。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recycle-bin/objects/restore": {
      "post": {
        "tags": ["回收站"],
        "operationId": "restoreRecycleBinObjects",
        "summary": "恢复回收站条目",
        "description": "一次最多恢复 200 个逻辑回收站条目。重复 ID 只处理一次，单项失败会返回在成功响应结果中。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecycleBinBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "包含 restored_count、failed_count 和 failed_items 的批量处理结果。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求体或 item_ids 列表无效。错误代码：invalid_request。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recycle-bin/objects/batch-delete": {
      "post": {
        "tags": ["回收站"],
        "operationId": "deleteRecycleBinObjects",
        "summary": "永久删除回收站条目",
        "description": "一次最多永久删除 200 个逻辑回收站条目。重复 ID 只处理一次，单项失败会返回在成功响应结果中。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecycleBinBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "包含 deleted_count、failed_count 和 failed_items 的批量处理结果。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "请求体或 item_ids 列表无效。错误代码：invalid_request。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Bearer Token 缺失或无效。错误代码：unauthorized。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "BucketParam": {
        "name": "bucket",
        "in": "path",
        "required": true,
        "description": "存储桶名称。必须匹配 `^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$`。",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$",
          "minLength": 3,
          "maxLength": 63
        }
      },
...
      "PrefixQuery": {
        "name": "prefix",
        "in": "query",
        "required": false,
        "description": "可选的前缀过滤器。最大长度： 512。文件夹样式前缀通常应以“/”结尾。",
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "required": false,
        "description": "应用于直接条目名称的不区分大小写的子字符串筛选器。",
        "schema": {
          "type": "string"
        }
      },
...
      "ExplorerLimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "页面大小。默认值为100。最大值为200。",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 200,
          "default": 100
        }
      },
      "CursorQuery": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "上一页返回的不透明光标。",
        "schema": {
          "type": "string"
        }
      },
      "FolderPathQuery": {
        "name": "path",
        "in": "query",
        "required": true,
        "description": "要删除的文件夹路径。必须是以“/”结尾的非空文件夹路径。",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "RecursiveQuery": {
        "name": "recursive",
        "in": "query",
        "required": false,
        "description": "`true` 时递归删除文件夹；省略或为 `false` 时，只能删除空文件夹。",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
...
    "schemas": {
...
      "SuccessEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": {
            "type": "string",
            "description": "中间件返回的请求关联ID。"
          },
          "data": {
            "type": "object",
            "description": "接口特定的响应数据。"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "FolderCreateRequest": {
        "type": "object",
        "required": ["prefix", "name"],
        "properties": {
          "prefix": {
            "type": "string",
            "maxLength": 512,
            "description": "父文件夹前缀。对存储桶根目录使用空字符串。非空值必须以“/”结尾。"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "description": "文件夹名称。不得包含“/”、“\\\\”、NUL、“.”、“..”或“.light-oss-folder”。"
          }
        }
      },
...
      "FolderNodeEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/FolderNode"
              }
            }
          }
        ]
      },
      "FolderNodeListEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessEnvelope"
          },
          {
            "type": "object",
            "required": ["request_id", "data"],
            "properties": {
              "request_id": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/FolderNodeList"
              }
            }
          }
        ]
      },
...
    }
  }
}